orion-commons/requirements.txt
aws-cdk.core==1.108.0
aws_cdk.aws_codepipeline==1.108.0
aws_cdk.aws_codepipeline_actions==1.108.0
aws_cdk.aws_iam==1.108.0
aws_cdk.aws_ssm==1.108.0
aws_cdk.pipelines==1.108.0
aws_cdk.aws_lambda==1.108.0
aws_cdk.aws_lambda_event_sources==1.108.0
aws_cdk.aws_s3==1.108.0
aws_cdk.aws_dynamodb==1.108.0
aws_cdk.aws_events==1.108.0